dev.p2ptradehelper/requirements.txt
django-simple-captcha
requests
django
docutils
gevent
grequests
git+https://github.com/Hibios/bestchange_api@main
djangorestframework
django-redis
celery
aiohttp
selenium
webdriver_manager
pytest
coverage
ratelimit
ratelimiter
gunicorn
locust
psycopg2
sentry-sdk
flower
django-nextjs
python-dateutil
drf-api-logger
django_resized
